+++ .eslintrc
@@ -1,0 +1,13 @@
+{
+  "plugins": ["jest", "@typescript-eslint"],
+  "parser": "@typescript-eslint/parser",
+  "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+  "env": {
+    "node": true,
+    "jest/globals": true
+  },
+  "ignorePatterns": ["*.test.ts"],
+  "rules": {
+    "@typescript-eslint/no-explicit-any": "off"
+  }
+}
