+++ .husky/pre-commit
@@ -1,4 +1,5 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
+npx lint-staged
 
 pnpm build && git add -A .
