+++ .lintstagedrc
@@ -1,0 +1,3 @@
+{
+  "./src/**/*.{js,ts}": ["prettier --write", "eslint --max-warnings 0"]
+}
