+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "singleQuote": true,
+  "jsxSingleQuote": true,
+  "singleAttributePerLine": true,
+  "useTabs": false,
+  "bracketSameLine": true
+}
