+++ package.json
@@ -12,8 +12,17 @@
   "license": "ISC",
   "devDependencies": {
     "@types/node": "^20.14.7",
+    "@types/jest": "^29.5.2",
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "eslint": "^8.42.0",
+    "eslint-plugin-jest": "^27.2.1",
     "husky": "^8.0.0",
-    "typescript": "^5.5.2"
+    "jest": "^29.5.0",
+    "lint-staged": "^13.2.2",
+    "prettier": "^2.8.8",
+    "ts-jest": "^29.1.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
   },
   "dependencies": {
     "@actions/core": "^1.10.0",
